requirements.txt
boto3>=1.34.0
botocore>=1.34.0
psycopg2-binary>=2.9.9
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
python-dateutil>=2.8.2
upstox-python-sdk>=2.0.0
requests>=2.31.0